+++ src/Stack/SetupCmd.hs
@@ -46,7 +46,7 @@ setupParser = SetupCmdOpts
                       "The default is to install the version implied by the resolver.")))
     <*> OA.boolFlags False
             "reinstall"
-            "reinstalling GHC, even if available (implies no-system-ghc)"
+            "reinstalling GHC, even if available (incompatible with --system-ghc)"
             OA.idm
     <*> OA.boolFlags False
             "upgrade-cabal"